move from go dep to go modules
