+++ Makefile
@@ -38,8 +38,8 @@ dev-run-go:
 	$(eval $@_DB_HOST := $(shell docker inspect --format '{{ .NetworkSettings.IPAddress }}' $(container_postgres_dev)))
 	$(eval $@_REDIS_HOST := $(shell docker inspect --format '{{ .NetworkSettings.IPAddress }}' $(container_redis_dev)))
 	$(eval $@_API_ENV := -e DB_HOST=$($@_DB_HOST) -e DB_NAME=$(db_name_dev) -e DB_PASS=$(db_pass_dev) -e DB_USER=$(db_user_dev) -e DB_SSL_MODE=disable -e REDIS_HOST=$($@_REDIS_HOST) -e REDIS_PORT=6379)
-	docker run -p 3001:8080 $($@_API_ENV) -v $(shell pwd)/src:/go/src/github.com/tylergeery/trash_hunt/src/ --name $(container_tcp_server_dev) -d $(image_tcp_server_dev)
-	docker run -p 3000:8080 $($@_API_ENV) -v $(shell pwd)/src:/go/src/github.com/tylergeery/trash_hunt/src/ --name $(container_api_server_dev) -d $(image_api_server_dev)
+	docker run -p 3001:8080 $($@_API_ENV) -v $(shell pwd)/src:/go/src/ --name $(container_tcp_server_dev) -d $(image_tcp_server_dev)
+	docker run -p 3000:8080 $($@_API_ENV) -v $(shell pwd)/src:/go/src/ --name $(container_api_server_dev) -d $(image_api_server_dev)
 
 dev-clean: ## Remove local docker images
 	- docker rmi $(image_api_server_dev) $(image_tcp_server_dev) $(image_postgres_dev)
@@ -47,7 +47,7 @@ dev-clean: ## Remove local docker images
 dev-images: ## Make dev docker images
 	docker build -f $(image_path)/pg/Dockerfile --target dev -t $(image_postgres_dev) $(build_path)
 	docker build -f $(image_path)/go/Dockerfile --target tcp_server_dev -t $(image_tcp_server_dev) $(build_path)
-	docker build -f $(image_path)/go/Dockerfile --target http_server_dev -t $(image_api_server_dev) $(build_path)
+	docker build -f $(image_path)/go/Dockerfile --target api_server_dev -t $(image_api_server_dev) $(build_path)
 
 dev-kill:
 	- docker kill $(container_api_server_dev) $(container_tcp_server_dev) $(container_postgres_dev) $(container_redis_dev)
