--- src/Gopkg.lock
@@ -1,88 +1,0 @@
-# This file is autogenerated, do not edit; changes may be undone by the next 'dep ensure'.
-
-
-[[projects]]
-  digest = "1:6098222470fe0172157ce9bbef5d2200df4edde17ee649c5d6e48330e4afa4c6"
-  name = "github.com/dgrijalva/jwt-go"
-  packages = ["."]
-  pruneopts = ""
-  revision = "06ea1031745cb8b3dab3f6a236daf2b0aa468b7e"
-  version = "v3.2.0"
-
-[[projects]]
-  digest = "1:ba0f277fb137e3d5d2cdaf09b286bd275577210c96c1e4ac0c99ddb10cc71743"
-  name = "github.com/go-ozzo/ozzo-routing"
-  packages = [
-    ".",
-    "content",
-    "cors",
-  ]
-  pruneopts = ""
-  revision = "9dcabc145f0d565b5219e2fd72c3da04876980b7"
-  version = "v2.1.4"
-
-[[projects]]
-  branch = "master"
-  digest = "1:158b29fe6a378ed096d48927038f7ad6c84ef28bd07a388b0688e91aa53feba2"
-  name = "github.com/golang/gddo"
-  packages = ["httputil/header"]
-  pruneopts = ""
-  revision = "af0f2af80721261f4d211b2c9563f7b46b2aab06"
-
-[[projects]]
-  digest = "1:dcf8316121302735c0ac84e05f4686e3b34e284444435e9a206da48d8be18cb1"
-  name = "github.com/gomodule/redigo"
-  packages = [
-    "internal",
-    "redis",
-  ]
-  pruneopts = ""
-  revision = "9c11da706d9b7902c6da69c592f75637793fe121"
-  version = "v2.0.0"
-
-[[projects]]
-  digest = "1:f32f8b1d82e7072d00b7c2d85a6fea53f204bcd5e28c4790a3d32a0ac87d7602"
-  name = "github.com/goware/emailx"
-  packages = ["."]
-  pruneopts = ""
-  revision = "80a00e4677ff2f102e17b78107f255253e2032e5"
-  version = "v0.2.0"
-
-[[projects]]
-  digest = "1:a65d93f562c7d66d8db6f0c5df8ade208e0f42d9a9626924896697351ca161b3"
-  name = "github.com/lib/pq"
-  packages = [
-    ".",
-    "oid",
-    "scram",
-  ]
-  pruneopts = ""
-  revision = "bc6a3c0594130b1e34005880bc600b6d3f49fa7f"
-  version = "v1.1.1"
-
-[[projects]]
-  branch = "master"
-  digest = "1:aebca5befed107d5990fc1868bc52b6e5aff2ebbac27c6f9814259d79f3c575a"
-  name = "golang.org/x/crypto"
-  packages = [
-    "bcrypt",
-    "blowfish",
-  ]
-  pruneopts = ""
-  revision = "ea8f1a30c4438cc8b13f05538385ad8dc6049b43"
-
-[solve-meta]
-  analyzer-name = "dep"
-  analyzer-version = 1
-  input-imports = [
-    "github.com/dgrijalva/jwt-go",
-    "github.com/go-ozzo/ozzo-routing",
-    "github.com/go-ozzo/ozzo-routing/content",
-    "github.com/go-ozzo/ozzo-routing/cors",
-    "github.com/gomodule/redigo/redis",
-    "github.com/goware/emailx",
-    "github.com/lib/pq",
-    "golang.org/x/crypto/bcrypt",
-  ]
-  solver-name = "gps-cdcl"
-  solver-version = 1
