--- src/Gopkg.toml
@@ -1,46 +1,0 @@
-
-# Gopkg.toml example
-#
-# Refer to https://github.com/golang/dep/blob/master/docs/Gopkg.toml.md
-# for detailed Gopkg.toml documentation.
-#
-# required = ["github.com/user/thing/cmd/thing"]
-# ignored = ["github.com/user/project/pkgX", "bitbucket.org/user/project/pkgA/pkgY"]
-#
-# [[constraint]]
-#   name = "github.com/user/project"
-#   version = "1.0.0"
-#
-# [[constraint]]
-#   name = "github.com/user/project2"
-#   branch = "dev"
-#   source = "github.com/myfork/project2"
-#
-# [[override]]
-#  name = "github.com/x/y"
-#  version = "2.4.0"
-
-
-[[constraint]]
-  name = "github.com/dgrijalva/jwt-go"
-  version = "3.2.0"
-
-[[constraint]]
-  name = "github.com/go-ozzo/ozzo-routing"
-  version = "2.1.4"
-
-[[constraint]]
-  name = "github.com/gomodule/redigo"
-  version = "2.0.0"
-
-[[constraint]]
-  name = "github.com/goware/emailx"
-  version = "0.2.0"
-
-[[constraint]]
-  name = "github.com/lib/pq"
-  version = "1.1.1"
-
-[[constraint]]
-  branch = "master"
-  name = "golang.org/x/crypto"
