+++ src/api_server/controllers/player.go
@@ -5,10 +5,10 @@ import (
 	"strconv"
 
 	"github.com/go-ozzo/ozzo-routing"
-	"github.com/tylergeery/trash_hunt/src/api_server/requests"
-	"github.com/tylergeery/trash_hunt/src/api_server/responses"
-	"github.com/tylergeery/trash_hunt/src/auth"
-	"github.com/tylergeery/trash_hunt/src/game"
+	"github.com/tylergeery/trash_hunt/api_server/requests"
+	"github.com/tylergeery/trash_hunt/api_server/responses"
+	"github.com/tylergeery/trash_hunt/auth"
+	"github.com/tylergeery/trash_hunt/game"
 )
 
 // PlayerCreate - Register/Signup a player
