+++ src/api_server/main.go
@@ -3,7 +3,7 @@ package main
 import (
 	"net/http"
 
-	"github.com/tylergeery/trash_hunt/src/api_server/router"
+	"github.com/tylergeery/trash_hunt/api_server/router"
 )
 
 func main() {
