+++ src/api_server/middleware/middleware_test.go
@@ -6,8 +6,8 @@ import (
 	"testing"
 
 	"github.com/go-ozzo/ozzo-routing"
-	"github.com/tylergeery/trash_hunt/src/auth"
-	"github.com/tylergeery/trash_hunt/src/game"
+	"github.com/tylergeery/trash_hunt/auth"
+	"github.com/tylergeery/trash_hunt/game"
 )
 
 type TestWriter struct {
