+++ src/go.mod
@@ -1,0 +1,13 @@
+module github.com/tylergeery/trash_hunt
+
+require (
+	github.com/dgrijalva/jwt-go v3.2.0+incompatible
+	github.com/go-ozzo/ozzo-routing v2.1.4+incompatible
+	github.com/golang/gddo v0.0.0-20190419222130-af0f2af80721 // indirect
+	github.com/gomodule/redigo v2.0.0+incompatible
+	github.com/google/go-cmp v0.3.0 // indirect
+	github.com/goware/emailx v0.2.0
+	github.com/lib/pq v1.1.1
+	github.com/stretchr/testify v1.3.0 // indirect
+	golang.org/x/crypto v0.0.0-20190618222545-ea8f1a30c443
+)
